+++ KeepingYouAwake/Preferences/Advanced/KYAAdvancedPreferencesViewController.m
@@ -47,10 +47,6 @@
                                                     defaultsKey:KYAUserDefaultsKeyNotificationsEnabled
                             ]];
     
-    [preferences addObject:[[KYAPreference alloc] initWithTitle:NSLocalizedString(@"Allow the display to sleep (when connected to AC power)", @"Allow the display to sleep (when connected to AC power)")
-                                                    defaultsKey:KYAUserDefaultsKeyAllowDisplaySleep
-                            ]];
-    
     [preferences addObject:[[KYAPreference alloc] initWithTitle:NSLocalizedString(@"Disable menu bar icon highlight color", @"Disable menu bar icon highlight color")
                                                     defaultsKey:KYAUserDefaultsKeyMenuBarIconHighlightDisabled
                             ]];